Updating Recommendations to only expose public features.
