+++ Sources/BVSwift/BVCommon/Submission/BVSubmission.swift
@@ -127,51 +127,27 @@ extension BVSubmission: BVURLRequestableCacheable {
   }
 }
 
-// MARK: - BVSubmission: BVURLParameterableInternal
-extension BVSubmission: BVURLParameterableInternal {
-  
+// MARK: - BVSubmission: BVURLParameterable
+extension BVSubmission: BVURLParameterable {
   final internal var parameters: [BVURLParameter] {
     return paramsPriv
   }
   
-  final internal func add(
-    _ parameter: BVURLParameter, coalesce: Bool = false) {
-    
-    guard coalesce else {
-      if 0 == paramsPriv.filter({ $0 === parameter }).count {
-        paramsPriv.append(parameter)
-      }
-      return
-    }
-    
-    var coalesceList: [BVURLParameter] = []
-    var otherList: [BVURLParameter] = []
-    paramsPriv.forEach { (param: BVURLParameter) in
-      if param %% parameter {
-        coalesceList.append(param)
-      } else {
-        otherList.append(param)
-      }
-    }
-    
-    let coalesce: BVURLParameter =
-      coalesceList.reduce(parameter, +~)
-    otherList.append(coalesce)
-    
-    paramsPriv = otherList
-  }
-  
-  final internal func update(_ parameter: BVURLParameter) {
-    var paramsTemp: [BVURLParameter] =
-      paramsPriv.filter { $0 !%% parameter }
-    paramsTemp.append(parameter)
-    
-    paramsPriv = paramsTemp
+  final internal func set(_ parameters: [BVURLParameter]) {
+    paramsPriv = parameters
   }
 }
 
 // MARK: - BVSubmission: BVSubmissionActionableInternal
 extension BVSubmission: BVSubmissionActionableInternal {
+  var bvPath: String {
+    return box.bvPath
+  }
+  
+  var commonEndpoint: String {
+    return box.commonEndpoint
+  }
+  
   var preflightHandler: BVURLRequestablePreflightHandler? {
     get {
       return box.preflightHandler
