+++ Tests/BVSwiftTests/BVRecommendations/BVRecommendationsQueryTest.swift
@@ -83,15 +83,18 @@ class BVRecommendationsQueryTest: XCTestCase {
   func testProfileQueryConstruction() {
     
     let bvBrandId: String = "610ce8a1-644f-4f02-8b6e-8b198376aa9d"
-    let lookbackTime: Int = 10
-    guard let lookback =
-      Calendar.current.date(
-        byAdding: .second,
-        value: -1 * lookbackTime,
-        to: Date()) else {
-          XCTFail()
-          return
-    }
+    
+    /*
+     let lookbackTime: Int = 10
+     guard let lookback =
+     Calendar.current.date(
+     byAdding: .second,
+     value: -1 * lookbackTime,
+     to: Date()) else {
+     XCTFail()
+     return
+     }
+     */
     
     let usLocale: Locale = Locale(identifier: "en_US")
     
@@ -100,7 +103,7 @@ class BVRecommendationsQueryTest: XCTestCase {
         /// Config
         .configure(BVRecommendationsQueryTest.config)
         /// Average Rating
-        .field(.avgRating(3.333333333333333333))
+        .field(.averageRating(3.333333333333333333))
         /// Brand Id
         .field(.brandId(bvBrandId))
         /// Interest
@@ -108,11 +111,15 @@ class BVRecommendationsQueryTest: XCTestCase {
         /// Locale
         .field(.locale(usLocale))
         /// Lookback
-        .field(.lookback(lookback))
+        //.field(.lookback(lookback))
         /// Includes
         .field(.include(.interests))
+        .field(.include(.interests))
+        .field(.include(.categories))
         .field(.include(.categories))
         .field(.include(.brands))
+        .field(.include(.brands))
+        .field(.include(.recommendations))
         .field(.include(.recommendations))
         /// Preferred Category
         .field(.preferredCategory("sports_wear"))
@@ -122,13 +129,13 @@ class BVRecommendationsQueryTest: XCTestCase {
         .field(.purpose(.ads))
         /// Required Category
         .field(.requiredCategory("menswear"))
-        /// Strategies
-        .field(.strategy("foo"))
-        .field(.strategy("foo"))
-        .field(.strategy("bar"))
-        .field(.strategy("bar"))
-        .field(.strategy("baz"))
-        .field(.strategy("baz"))
+    /// Strategies
+    //.field(.strategy("foo"))
+    //.field(.strategy("foo"))
+    //.field(.strategy("bar"))
+    //.field(.strategy("bar"))
+    //.field(.strategy("baz"))
+    //.field(.strategy("baz"))
     
     _ = profileQuery.preflight { (error: Error?) in
       guard nil == error else {
@@ -154,13 +161,13 @@ class BVRecommendationsQueryTest: XCTestCase {
         "include=brands,category_recommendations,interests,recommendations"))
     XCTAssertTrue(requestString.contains("interest=football"))
     XCTAssertTrue(requestString.contains("locale=\(usLocale.identifier)"))
-    XCTAssertTrue(requestString.contains("lookback=10s"))
+    //XCTAssertTrue(requestString.contains("lookback=10s"))
     XCTAssertTrue(
       requestString.contains("product=apitestcustomer/product1234"))
     XCTAssertTrue(requestString.contains("purpose=ads"))
     XCTAssertTrue(
       requestString.contains("required_category=apitestcustomer/menswear"))
-    XCTAssertTrue(requestString.contains("strategies=bar,baz,foo"))
+    //XCTAssertTrue(requestString.contains("strategies=bar,baz,foo"))
   }
   
   func testProfileQuery() {
